+++ protected/humhub/modules/admin/models/GroupSearch.php
@@ -8,9 +8,11 @@
 
 namespace humhub\modules\admin\models;
 
+use humhub\modules\user\models\forms\EditGroupForm;
+use humhub\modules\user\models\Group;
 use yii\base\Model;
 use yii\data\ActiveDataProvider;
-use humhub\modules\user\models\Group;
+use yii\db\Expression;
 
 /**
  * Description of GroupSearch
@@ -19,10 +21,12 @@ use humhub\modules\user\models\Group;
  */
 class GroupSearch extends Group
 {
+    public $type;
+
     public function rules()
     {
         return [
-            [['name', 'description'], 'safe'],
+            [['name', 'description', 'type'], 'safe'],
         ];
     }
 
@@ -44,7 +48,7 @@ class GroupSearch extends Group
      */
     public function search($params)
     {
-        $query = Group::find()->orderBy(['sort_order' => SORT_ASC, 'name' => SORT_ASC]);
+        $query = Group::find();
 
         $dataProvider = new ActiveDataProvider([
             'query' => $query,
@@ -55,6 +59,13 @@ class GroupSearch extends Group
             'attributes' => [
                 'name',
                 'descriptions',
+                'type' => [
+                    'asc' => ['`parent_group_id` IS NULL' => SORT_ASC, 'name' => SORT_ASC],
+                    'desc' => ['`parent_group_id` IS NULL' => SORT_DESC, 'name' => SORT_ASC],
+                ],
+            ],
+            'defaultOrder' => [
+                'type' => SORT_DESC,
             ],
         ]);
 
@@ -67,6 +78,11 @@ class GroupSearch extends Group
 
         $query->andFilterWhere(['like', 'name', $this->name]);
         $query->andFilterWhere(['like', 'description', $this->description]);
+
+        if (!empty($this->type)) {
+            $operator = $this->type === EditGroupForm::TYPE_NORMAL ? 'IS' : 'IS NOT';
+            $query->andFilterWhere([$operator, 'parent_group_id', new Expression('NULL')]);
+        }
 
         return $dataProvider;
     }